backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
albumentations>=1.3.0
kaggle>=1.5.16
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gradio>=3.44.0
plotly>=5.15.0
tqdm>=4.65.0
requests>=2.31.0
